requirements.txt
asyncpg==0.30.0
pyotp==2.9.0
numpy==2.2.4
scipy==1.15.2
pandas==2.2.3
pyarrow==19.0.1
aiohttp==3.11.14
websockets==14.2
python-decouple==3.8
APScheduler==3.11.0
requests==2.32.3
loguru==0.7.3
python-telegram-bot==21.10
httpx==0.27.2
yfinance==0.2.54
psycopg2-binary==2.9.10
lxml==5.3.1